Revert "move load_params() from container to policy"
This reverts commit be50a6487db3de2bd0ad6b6363418fccdb4c865b.
