+++ deepspeed/module_inject/containers/gptneo.py
@@ -21,45 +21,6 @@ class DS_GPTNEOContainer(MetaTensorContainer, BaseTransformerContainer):
         self.module.config.scale_attention = self.scale_attention
         return self.module
 
-
-class HFGPTNEOLayerPolicy(TransformerPolicy):
-    def __init__(self, client_module, inference=True):
-        super().__init__(inference, scale_attention=False)
-        self.client_module = client_module
-        try:
-            import transformers
-            HFGPTNEOLayerPolicy._orig_layer_class = transformers.models.gpt_neo.modeling_gpt_neo.GPTNeoBlock
-        except:
-            HFGPTNEOLayerPolicy._orig_layer_class = None
-
-    def get_hidden_heads(self):
-        return self.client_module.attn.attention.q_proj.weight.shape[1], \
-                self.client_module.attn.attention.num_heads
-
-    def attention(self):
-        qw = self.client_module.attn.attention.q_proj.weight
-        kw = self.client_module.attn.attention.k_proj.weight
-        vw = self.client_module.attn.attention.v_proj.weight
-
-        qkvw = Parameter(torch.cat((qw, kw, vw), dim=0), requires_grad=False)
-
-        return qkvw, \
-               None, \
-               self.client_module.attn.attention.out_proj.weight, \
-               self.client_module.attn.attention.out_proj.bias
-
-    def mlp(self):
-        return self.client_module.mlp.c_fc.weight, \
-               self.client_module.mlp.c_fc.bias, \
-               self.client_module.mlp.c_proj.weight, \
-               self.client_module.mlp.c_proj.bias
-
-    def layernorm(self):
-        return self.client_module.ln_2.weight, \
-               self.client_module.ln_2.bias, \
-               self.client_module.ln_1.weight, \
-               self.client_module.ln_1.bias
-
     def load_params(self, module, sd, weight_quantizer, mp_replace, prefix):
         param_names = (
             'attn.attention.q_proj.weight', \
@@ -107,3 +68,42 @@ class HFGPTNEOLayerPolicy(TransformerPolicy):
                        mp_replace,
                        transformer_param_names[i - 1],
                        prefix + param_names[i])
+
+
+class HFGPTNEOLayerPolicy(TransformerPolicy):
+    def __init__(self, client_module, inference=True):
+        super().__init__(inference, scale_attention=False)
+        self.client_module = client_module
+        try:
+            import transformers
+            HFGPTNEOLayerPolicy._orig_layer_class = transformers.models.gpt_neo.modeling_gpt_neo.GPTNeoBlock
+        except:
+            HFGPTNEOLayerPolicy._orig_layer_class = None
+
+    def get_hidden_heads(self):
+        return self.client_module.attn.attention.q_proj.weight.shape[1], \
+                self.client_module.attn.attention.num_heads
+
+    def attention(self):
+        qw = self.client_module.attn.attention.q_proj.weight
+        kw = self.client_module.attn.attention.k_proj.weight
+        vw = self.client_module.attn.attention.v_proj.weight
+
+        qkvw = Parameter(torch.cat((qw, kw, vw), dim=0), requires_grad=False)
+
+        return qkvw, \
+               None, \
+               self.client_module.attn.attention.out_proj.weight, \
+               self.client_module.attn.attention.out_proj.bias
+
+    def mlp(self):
+        return self.client_module.mlp.c_fc.weight, \
+               self.client_module.mlp.c_fc.bias, \
+               self.client_module.mlp.c_proj.weight, \
+               self.client_module.mlp.c_proj.bias
+
+    def layernorm(self):
+        return self.client_module.ln_2.weight, \
+               self.client_module.ln_2.bias, \
+               self.client_module.ln_1.weight, \
+               self.client_module.ln_1.bias
